Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,7 +2,7 @@
 int main(){
     int x;
     scanf("%d",&x);
-    int Area = 3.14*x*x;
-    printf("%d",Area);
+    float Area = 3.14*x*x;
+    printf("%.2f",Area);
     return 0;
 }
